completed the split of write service and rightsholder responsibilities
